add first version of hovers for static scripting variable references
